Fix/#7307 numeric to bigint

+++ src/entities/monitor-location.entity.ts
@@ -21,7 +21,7 @@ import {
   
     @Column({
       name: 'unit_id',
-      type: 'numeric',
+      type: 'bigint',
       transformer: new NumericColumnTransformer(),
     })
     unitIdentifier: number;
